app: remove Arc<Tx> from ActionHandler impls
This is done by:

- making a new `TransactionContext` that has only the context necessary for verification, the anchor and the effect hash
- putting it at the bottom of the dep tree, in the crypto crate
- threading it through the new `ActionHandler` associated types.
