+++ crypto/src/dex/swap/plaintext.rs
@@ -2,8 +2,8 @@ use crate::address::AddressVar;
 use crate::asset::{AmountVar, AssetIdVar};
 use crate::keys::OutgoingViewingKey;
 use crate::note::StateCommitmentVar;
-use crate::transaction::Fee;
 use crate::value::ValueVar;
+use crate::Fee;
 use crate::{asset, ka, Address, Amount, Note, PayloadKey, Rseed, Value};
 use anyhow::{anyhow, Error, Result};
 use ark_ff::PrimeField;
